// lib/types.ts
export type CustomUser = NextAuthUser & Omit<PrismaUser, 'createdAt'> & {
  createdAt?: Date;
  loginPortal?: "ADMIN" | "VISITOR";
};

export interface CreateUserEmailProps {
  user: {
    name: string | null | undefined;
    email: string | null | undefined;
  };
}